created RelatedItems and OutfitList components and refactored code to have RelatedItems rendered to App and have RelatedItems contain the Carousel component
