+++ client/src/components/App.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import QandA from './QandA/QandA.jsx';
 import AllReviews from './RatingsReview/AllReviews.jsx';
 import ProductInfo from './Overview/SM-ProductInfo.jsx';
-import Carousel from './RelatedItemsAndOutfit/Carousel.jsx';
+import RelatedItems from './RelatedItemsAndOutfit/RelatedItems.jsx';
 import RatingBreakdown from './RatingsReview/RatingBreakdown.jsx';
 
 class App extends React.Component {
@@ -14,13 +14,13 @@ class App extends React.Component {
     return (
       <>
       <RatingBreakdown />
-        {/* <h1>Kloth</h1>
+        <h1>Kloth</h1>
         <ProductInfo />
-        <Carousel renderedId={'42368'}/>
+        <RelatedItems renderedId={'42368'}/>
         <div>
           <QandA />
         </div>
-        <AllReviews /> */}
+        <AllReviews />
       </>
     );
   }
